Implement toy collaborative fitlering model
