+++ src/toymodels/content_based_model.py
@@ -1,4 +1,21 @@
-from movie import Movie
+from dataclasses import dataclass
+
+
+@dataclass(frozen=True)
+class Movie:
+    name: str
+    animated: bool
+    is_marvel: bool
+    has_super_villain: bool
+    passes_bechdel_test: bool
+
+    def match(self, other):
+        score = 0
+        for attr, value in self.__dict__.items():
+            if value == getattr(other, attr):
+                score += 1
+
+        return score
 
 
 def find_best_match(movies, similar_to):
